pyproject.toml
[project]
name = "langgraph_learning"
version = "0.1.0"
description = "A learning project for langgraph"
readme = "README.md"
authors = [{ name = "lostyear", email = "lostyear540@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-text-splitters>=1.1.0",
    "markdown>=3.10",
    "python-dotenv>=1.2.1",
    "unstructured>=0.18.27",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
]
